+++ go/basic_item.go
@@ -3,10 +3,10 @@ package codingdojo
 type BasicItem struct {
 	name           string
 	baseDamage     int32
-	damageModifier float32
+	damageModifier float64
 }
 
-func makeBasicItem(name string, baseDamage int32, damageModifier float32) BasicItem {
+func makeBasicItem(name string, baseDamage int32, damageModifier float64) BasicItem {
 	return BasicItem{name, baseDamage, damageModifier}
 }
 
@@ -14,6 +14,6 @@ func (this BasicItem) getBaseDamage() int32 {
 	return this.baseDamage
 }
 
-func (this BasicItem) getDamageModifier() float32 {
+func (this BasicItem) getDamageModifier() float64 {
 	return this.damageModifier
 }
